requirements.txt
pytorch-lightning==2.5.6
pyloudnorm==0.1.0
deep-phonemizer==0.0.17
soundfile==0.10.3.post1
tqdm==4.62.3
torch==2.8.0
torchaudio==2.8.0
pyannote.audio==3.1.1
numpy==1.26.4
librosa==0.11.0
huggingface_hub==0.15.0
transformers==4.57.1
